src/kernel/pyproject.toml
[project]
name = "mustang-kernel"
dynamic = ["version"]
description = "A fully customized modularize AI agent kernel"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Haolei (Saki) Ye" },
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "httpx>=0.28",
    "anthropic[bedrock]>=0.50",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "websockets>=13.0",
    "html2text>=2024.2",
    "croniter>=1.3",
    "jieba>=0.42",
    "orjson>=3.10",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.24"]
web-readability = ["readability-lxml>=0.8", "lxml>=5.0"]
web-browser = ["playwright>=1.40"]
web-all = ["readability-lxml>=0.8", "lxml>=5.0", "playwright>=1.40"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "kernel/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["kernel"]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "httpx[asyncio]>=0.28.1",
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "pytest-anyio>=0.0.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.anyio]
backend = "asyncio"

[project.scripts]
mustang-kernel = "kernel.__main__:main"
